requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
opencv-python==4.9.0.80
numpy<2.0
boto3==1.34.51
botocore==1.34.51
Flask-CORS==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0 